Split monomial
