+++ src/CMakeLists.txt
@@ -297,6 +297,8 @@ libcvc4_add_sources(
   theory/arith/nl_lemma_utils.h
   theory/arith/nl_model.cpp
   theory/arith/nl_model.h
+  theory/arith/nl_monomial.cpp
+  theory/arith/nl_monomial.h
   theory/arith/nl_solver.cpp
   theory/arith/nl_solver.h
   theory/arith/nonlinear_extension.cpp
