+++ src/theory/arith/nonlinear_extension.cpp
@@ -17,17 +17,10 @@
 
 #include "theory/arith/nonlinear_extension.h"
 
-#include <cmath>
-#include <set>
-
-#include "expr/node_algorithm.h"
-#include "expr/node_builder.h"
 #include "options/arith_options.h"
-#include "theory/arith/arith_msum.h"
-#include "theory/arith/arith_utilities.h"
 #include "theory/arith/theory_arith.h"
 #include "theory/ext_theory.h"
-#include "theory/quantifiers/quant_util.h"
+#include "theory/arith/arith_utilities.h"
 #include "theory/theory_model.h"
 
 using namespace CVC4::kind;
